server/python-flask/requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
connexion==2.14.2
cryptography==41.0.3
ecdsa==0.18.0
Flask==2.2.5
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
swagger-ui-bundle==0.0.9
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
